Back-fill extra meta-faucets for MORE FAUCETS upgrade.
Purchasing the MORE FAUCETS upgrade is supposed to give you an extra
Meta-Faucet for each perfect square of Meta-Faucet Upgrades.
Unfortunately, this wasn't applied retroactively for previously
purchased Meta-Faucet Upgrades, and since at this point they're
auto-purchased it meant that players had to race to buy this upgrade
before the first Meta-Faucet Upgrade was bought. For some people this is
nearly impossible to do given how quickly the auto-purchasing happens,
and without the extra Meta-Faucets you're effectively stuck.